Map functionality updated
The map (ZoomablePaneTest.java) has been updated so it can now be scaled and moved with any amount of nodes in it. It also clips and stays within a defined area (black box). Units can be added to the map using a button, can be selected by clicking them and can be moved using WASD. Units are still slowed when walking through water (altho the "selected unit effect" slightly mismatches the pixel which is read).